Move follow request's settings updating to following controller
