+++ backend/src/Civix/ApiBundle/Controller/V2/UserFollowerController.php
@@ -2,11 +2,9 @@
 
 namespace Civix\ApiBundle\Controller\V2;
 
-use Civix\ApiBundle\Form\Type\UserFollowType;
 use Civix\CoreBundle\Entity\UserFollow;
 use Civix\CoreBundle\Service\FollowerManager;
 use Doctrine\ORM\EntityManagerInterface;
-use FOS\RestBundle\Controller\Annotations as REST;
 use FOS\RestBundle\Controller\Annotations\QueryParam;
 use FOS\RestBundle\Controller\Annotations\View;
 use FOS\RestBundle\Controller\FOSRestController;
@@ -151,56 +149,6 @@ class UserFollowerController extends FOSRestController
     public function patchAction(UserFollow $userFollow): void
     {
         $this->manager->approve($userFollow);
-    }
-
-    /**
-     * Edit a follow request
-     *
-     * @REST\Put("/{id}", requirements={"id"="\d+"})
-     *
-     * @ParamConverter("userFollow", options={"mapping" = {"loggedInUser" = "user", "id" = "follower"}}, converter="doctrine.param_converter")
-     *
-     * @ApiDoc(
-     *     authentication=true,
-     *     section="Followers",
-     *     requirements={
-     *         {
-     *             "name"="id",
-     *             "dataType"="integer",
-     *             "requirement"="\d+",
-     *             "description"="User id"
-     *         }
-     *     },
-     *     input="Civix\ApiBundle\Form\Type\UserFollowType",
-     *     output={
-     *          "class" = "Civix\CoreBundle\Entity\UserFollow",
-     *          "groups" = {"api-info", "api-followers"},
-     *          "parsers" = {
-     *              "Nelmio\ApiDocBundle\Parser\JmsMetadataParser"
-     *          }
-     *     },
-     *     statusCodes={
-     *         404="Follow Request Not Found"
-     *     }
-     * )
-     *
-     * @View(serializerGroups={"api-info", "api-followers"})
-     *
-     * @param Request $request
-     * @param UserFollow $userFollow
-     *
-     * @return UserFollow|\Symfony\Component\Form\Form
-     */
-    public function putAction(Request $request, UserFollow $userFollow)
-    {
-        $form = $this->createForm(UserFollowType::class, $userFollow);
-        $form->submit($request->request->all());
-
-        if ($form->isValid()) {
-            return $this->manager->save($userFollow);
-        }
-
-        return $form;
     }
 
     /**
